feat: re-add alarm clock sound when time end
