+++ src/app/layout.tsx
@@ -1,7 +1,5 @@
-import { TimerProvider } from '@/context/timer.provider';
 import './globals.css'
 import type { Metadata } from "next";
-import { ThemeProvider } from '@/context/theme.provider';
 
 export const metadata: Metadata = {
   title: 'My Pomodoro',
@@ -17,11 +15,7 @@ export default function RootLayout({
     <html lang="en">
       <link rel="shortcut icon" type="image/svg" href="/assets/images/clock.svg" />
       <body>
-        <ThemeProvider>
-          <TimerProvider>
-            {children}
-          </TimerProvider>
-        </ThemeProvider>
+        {children}
       </body>
     </html>
   );
