+++ src/context/index.tsx
@@ -1,0 +1,5 @@
+export type { ThemeContextType } from "./theme.provider";
+export { ThemeContext, ThemeProvider } from "./theme.provider";
+
+export type { TimerContextType } from "./timer.provider";
+export { TimerContext, TimerProvider } from "./timer.provider";
